feat:Add upload NO_CERT event

+++ sysapp/agent/core/bootstrap/src/profile_parse.c
@@ -89,22 +89,22 @@ static uint16_t rt_init_file_info(rt_fshandle_t fp)
     MSG_PRINTF(LOG_INFO, "operator_num:%d\n", data.operator_num);
 }
 
-static uint32_t rt_get_root_sk_offset(rt_fshandle_t fp, uint8_t *sk, uint16_t *size)
+static uint32_t rt_get_root_sk_offset(rt_fshandle_t fp, uint8_t *sk, uint32_t *size)
 {
     return get_offset(fp, ROOT_SK, sk, size);
 }
 
-static uint32_t rt_get_aes_key_offset(rt_fshandle_t fp, uint8_t *sk, uint16_t *size)
+static uint32_t rt_get_aes_key_offset(rt_fshandle_t fp, uint8_t *sk, uint32_t *size)
 {
     return get_offset(fp, PROFILE_KEY, sk, size);
 }
 
-static uint32_t rt_get_file_info_offset(rt_fshandle_t fp, uint8_t *sk, uint16_t *size)
+static uint32_t rt_get_file_info_offset(rt_fshandle_t fp, uint8_t *sk, uint32_t *size)
 {
     return get_offset(fp, FILE_INFO, sk, size);
 }
 
-static uint32_t rt_get_operator_profile_offset(rt_fshandle_t fp, uint8_t *sk, uint16_t *size)
+static uint32_t rt_get_operator_profile_offset(rt_fshandle_t fp, uint8_t *sk, uint32_t *size)
 {
     return get_offset(fp, OPT_PROFILES, sk, size);
 }
@@ -206,8 +206,7 @@ static int32_t update_hash(uint8_t *buf, int32_t profile_len, uint8_t *profile_h
     return RT_SUCCESS;
 }
 
-static int32_t
-build_profile(uint8_t *profile_buffer, int32_t profile_len, int32_t selected_profile_index, BOOLEAN_t sequential)
+static int32_t build_profile(uint8_t *profile_buffer, int32_t profile_len, int32_t selected_profile_index, BOOLEAN_t sequential)
 {
     BootstrapRequest_t *bootstrap_request = NULL;
     asn_dec_rval_t dc;
@@ -271,7 +270,7 @@ build_profile(uint8_t *profile_buffer, int32_t profile_len, int32_t selected_pro
     MSG_INFO_ARRAY("Current profile:", g_buf, g_buf_size);
     msg_send_agent_queue(MSG_ID_CARD_MANAGER, MSG_CARD_SETTING_PROFILE, g_buf, g_buf_size);
     ret = RT_SUCCESS;
-    end:
+end:
     rt_os_free(g_buf);
     g_buf = NULL;
     return ret;
@@ -379,7 +378,7 @@ int32_t init_profile_file(int32_t *arg)
 {
     int32_t ret = RT_SUCCESS;
     uint8_t buf[500];
-    uint16_t len = 0;
+    uint32_t len = 0;
     rt_fshandle_t fp;
 
     fp = rt_fopen(SHARE_PROFILE, RT_FS_READ);
